Clean up | Handling of non existing files

+++ src/file_handling.rs
@@ -1,7 +1,4 @@
-use std::{
-    fs, io,
-    path::{Path, PathBuf},
-};
+use std::{fs, io, path::Path};
 
 use crate::server::send_response;
 
@@ -9,11 +6,14 @@ const CREATE_DIR_INSTR: &str = "CRTDIR";
 const CREATE_FILE_INSTR: &str = "CRTFILE";
 const DELETE_FILE_INSTR: &str = "DELFILE";
 const DELETE_DIR_INSTR: &str = "DELDIR";
+const READ_DIR_INSTR: &str = "READDIR";
+const READ_FILE_INSTR: &str = "READFILE";
 
 #[derive(PartialEq, Debug)]
 enum FileResult {
-    CREATED,
-    EXISTS,
+    Success,
+    Exists,
+    DoesNotExist,
 }
 
 fn dir_exists(dir: &str) -> bool {
@@ -38,63 +38,44 @@ pub fn is_inside_root(path: &String) -> bool {
 
 fn create_path_to(path: &str) -> std::io::Result<FileResult> {
     if dir_exists(&path) {
-        return Ok(FileResult::EXISTS);
+        return Ok(FileResult::Exists);
     }
 
     fs::create_dir_all(path)?;
 
-    Ok(FileResult::CREATED)
+    Ok(FileResult::Success)
 }
 
-fn create_file_at<'a>(path: &'a str) -> std::io::Result<FileResult> {
+fn create_file_at(path: &str) -> std::io::Result<FileResult> {
     if file_exists(&path) {
-        return Ok(FileResult::EXISTS);
+        return Ok(FileResult::Exists);
     }
 
     fs::File::create(&path)?;
 
-    Ok(FileResult::CREATED)
+    Ok(FileResult::Success)
 }
 
-fn delete_file(path: &Path) -> std::io::Result<()> {
-    fs::remove_file(path)?;
-    Ok(())
+fn delete_file(path: &str) -> std::io::Result<FileResult> {
+    if !file_exists(&path) {
+        return Ok(FileResult::DoesNotExist);
+    }
+    fs::remove_file(&path)?;
+    Ok(FileResult::Success)
 }
 
-fn delete_dir(path: &Path) -> std::io::Result<()> {
+fn delete_dir(path: &str) -> std::io::Result<FileResult> {
+    if !dir_exists(&path) {
+        return Ok(FileResult::DoesNotExist);
+    }
     fs::remove_dir_all(path)?;
-    Ok(())
+    Ok(FileResult::Success)
 }
 
 pub fn init_root() {
-    let dir_entries = fs::read_dir(".");
-
-    match dir_entries {
-        Ok(entries) => {
-            let entry_vec: Result<Vec<PathBuf>, std::io::Error> =
-                entries.map(|res| res.map(|e| e.path())).collect();
-
-            match entry_vec {
-                Ok(entr_vec) => {
-                    if entr_vec.contains(&PathBuf::from("./root")) {
-                        return;
-                    }
-
-                    match fs::create_dir("./root") {
-                        Ok(_) => {}
-                        Err(_) => {
-                            panic!("Could not create root | Aborting")
-                        }
-                    }
-                }
-                Err(_) => {
-                    //can not happen
-                }
-            }
-        }
-        Err(_) => {
-            panic!("Directory could not be read | Aborting")
-        }
+    match fs::create_dir("./root") {
+        Ok(_) => {}
+        Err(_) => {}
     }
 }
 
@@ -103,50 +84,39 @@ pub async fn execute_instruction(
     path: &str,
     socket: &mut tokio::net::TcpStream,
 ) -> io::Result<()> {
-    //TODO make this more compact - helper function - maybe an enum
     let response: &str = match instr {
         CREATE_DIR_INSTR => match create_path_to(path) {
-            Ok(file_res) => {
-                if file_res == FileResult::CREATED {
-                    "HTTP/1.1 200 OK\r\n\r\nCreated directory!";
-                }
-                if file_res == FileResult::EXISTS {
-                    "HTTP/1.1 200 OK\r\n\r\nDirectory already exists!"
-                } else {
-                    //TODO this is reached sometimes
-                    "HTTP/1.1 400 Bad\r\n\r\nNot reachable!"
-                }
-            }
+            Ok(file_res) => match file_res {
+                FileResult::Success => "HTTP/1.1 200 OK\r\n\r\nCreated directory!",
+                FileResult::DoesNotExist => "HTTP/1.1 200 OK\r\n\r\nUnreachable",
+                FileResult::Exists => "HTTP/1.1 200 OK\r\n\r\nDirectory already exists!",
+            },
             Err(_) => "HTTP/1.1 400 Bad\r\n\r\nCargo was valid. Failed to create directory!",
         },
         CREATE_FILE_INSTR => match create_file_at(path) {
-            Ok(file_res) => {
-                if file_res == FileResult::CREATED {
-                    "HTTP/1.1 200 OK\r\n\r\nCreated file!";
-                }
-                if file_res == FileResult::EXISTS {
-                    "HTTP/1.1 200 OK\r\n\r\nFile already exists!"
-                } else {
-                    //TODO this is reached sometimes
-                    "HTTP/1.1 400 Bad\r\n\r\nNot reachable!"
-                }
-            }
+            Ok(file_res) => match file_res {
+                FileResult::Success => "HTTP/1.1 200 OK\r\n\r\nCreated file!",
+                FileResult::DoesNotExist => "HTTP/1.1 200 OK\r\n\r\nUnreachable",
+                FileResult::Exists => "HTTP/1.1 200 OK\r\n\r\nFile already exists!",
+            },
             Err(_) => "HTTP/1.1 400 Bad\r\n\r\nCargo was valid. Failed to create file!",
         },
-        DELETE_FILE_INSTR => {
-            delete_file(Path::new(path))?;
-
-            //TODO handle case if file does not exist
-
-            "HTTP/1.1 200 OK\r\n\r\nDeleted File!"
-        }
-        DELETE_DIR_INSTR => {
-            delete_dir(Path::new(path))?;
-
-            //TODO handle case if dir does not exist
-
-            "HTTP/1.1 200 OK\r\n\r\nDeleted Dir"
-        }
+        DELETE_FILE_INSTR => match delete_file(path) {
+            Ok(file_res) => match file_res {
+                FileResult::Success => "HTTP/1.1 200 OK\r\n\r\nDeleted file!",
+                FileResult::DoesNotExist => "HTTP/1.1 200 OK\r\n\r\nFile does not exist!",
+                FileResult::Exists => "HTTP/1.1 200 OK\r\n\r\nUnreachable",
+            },
+            Err(_) => "HTTP/1.1 400 Bad\r\n\r\nPayload was valid. Failed to delete file!",
+        },
+        DELETE_DIR_INSTR => match delete_dir(path) {
+            Ok(file_res) => match file_res {
+                FileResult::Success => "HTTP/1.1 200 OK\r\n\r\nDeleted directory!",
+                FileResult::DoesNotExist => "HTTP/1.1 200 OK\r\n\r\nDirectory does not exist!",
+                FileResult::Exists => "HTTP/1.1 200 OK\r\n\r\nUnreachable",
+            },
+            Err(_) => "HTTP/1.1 400 Bad\r\n\r\nPayload was valid. Failed to delete file!",
+        },
         &_ => "HTTP/1.1 400 Bad\r\n\r\nUnknown instruction\r\n\r\n",
     };
